project/requirements.txt
asgiref==3.7.2
dj-database-url==2.3.0
Django==4.2.11
djangorestframework==3.16.0
pillow==11.1.0
sqlparse==0.4.4
typing_extensions==4.10.0